6.3. Added UserBlock component.
